binance-producer/requirements.txt
kafka-python
python-binance

